Fixing graceful shutdown
From
https://ci--alfred.elifesciences.org/job/test-elife-lax/110/consoleFull
```
      ID: bot-lax-adaptors-start
Function: cmd.run
    Name: start bot-lax-adaptors
  Result: True
 Comment: Command "start

 Started: 11:52:22.340042
Duration: 30129.05 ms
 Changes:
          ----------
          pid:
              3412
          retcode:
              0
          stderr:
          stdout:
              bot-lax-adaptors stop/waiting
```

The graceful shutdown times out after 30 seconds and the process gets
killed with `SIGKILL`.

The reason is if there are no messages (which is 99% of the time), the
inner while loop returns to polling without yielding a new value,
never exiting.

Therefore, passing the flag to the IncomingQueue allows it to check
whether it should stop, before starting any new polling. If polling
returned something, that is always processed before attempting a new
iteration.

+++ src/sqs_adaptor.py
@@ -26,21 +26,25 @@ def poll(queue_obj):
         message.delete()
 
 class IncomingQueue(object):
-    def __init__(self, queue_name):
+    def __init__(self, queue_name, flag=None):
         self.queue = conn().get_queue_by_name(QueueName=queue_name)
+        self.flag = flag
 
     def __iter__(self):
         """an infinite poll on the given queue object.
         blocks for 20 seconds before connection is dropped and re-established"""
         while True:
-            messages = []
-            while not messages:
-                messages = self.queue.receive_messages(
-                    MaxNumberOfMessages=1,
-                    VisibilityTimeout=60, # time allowed to call delete, can be increased
-                    WaitTimeSeconds=20 # maximum setting for long polling
-                )
-            LOG.debug("processing sqs message")
+            if self.flag.should_stop:
+                break
+            messages = self.queue.receive_messages(
+                MaxNumberOfMessages=1,
+                VisibilityTimeout=60, # time allowed to call delete, can be increased
+                WaitTimeSeconds=20 # maximum setting for long polling
+            )
+            if not messages:
+                continue
+
+            LOG.debug("processing message")
             message = messages[0]
             yield message.body
             message.delete()
